Array macht es übersichtlicher /shellcheck
shellcheck prüft und fordert, dass ich die sender arrays quotieren soll. Ich kriegs net gelöst. Wenn ich das mache, dann funktioniert das Script nicht mehr.  Irgend`welche Ideen?

wahl=$(kdialog --title "K+Radio" --radiolist "$station" ${sender1[*]} ${sender2[*]} ${sender3[*]})

+++ Auto_menu_Auswahl.sh
@@ -1,57 +1,70 @@
 #!/bin/bash
 
+# Dieses Radio braucht eine Liste
+
+## entweder von Github downloaden
+#radio.txt in den selben Ordner kopieren wie das Script
+
+## oder statt "radio.txt" in der Zeile mpg123 …
+# diese URL:
+# https://raw.githubusercontent.com/dewomser/Webradio-Bash-Dialog/master/radio.txt
+#eintippen
+
+
+command -v kdialog >/dev/null 2>&1 || { echo -e "\e[41mDieses Radio benötigt Kdialog. Das Programm 'kdialog' muss installiert werden.\e[0m" >&2; exit 1; }
+command -v mpg123 >/dev/null 2>&1 || { kdialog --sorry "Kdialog ist installiert, wenn Du jetzt noch\n'mpg123' instalierst, funktioniert das Radio !" >&2; exit 1; }
+
+
+sender1=(1 DotFM8072 off)
+sender2=(2 DotFM8042 off)
+sender3=(3 SWR3 off)
+ 
+
+loopy=1
 rm  fifo.txt
 touch fifo.txt
-station_status_3="SWR_3 off"
-station_status_2="Dot_FM_8042 off"
-station_status_1="Dot_FM_8072 off"
 
 radiostation()          {
-                      mpg123 --control --utf8 --title --preload 1 --buffer 768 --smooth "$1" >> fifo.txt  2>&1> /dev/null &
+                      #echo"">fifo.txt
+                      mpg123 --control --utf8 -@ radio.txt --title --preload 1 --buffer 768 --smooth -l "$1" >> fifo.txt  2>&1> /dev/null &
+                      sleep 1
                       result=$(tail -n 25 fifo.txt|grep -a --line-buffered "StreamTitle"| sed -e 's/;.*//' -e 's/.*=//' -e "s/'//g")
-                      station=$1
+                      station=$(tail -n 10 fifo.txt|grep -m1 --line-buffered "ICY-NAME"| sed -e 's/ICY-NAME: //' -e 's/ /_/g' -e 's/\./_/g' )
                       kdialog --title "Radio Info" --passivepopup "$result" 10
                       
                       }
-                      
+while [ "$loopy" -eq 1 ] ; do
         
         
-        
-       
-    while true; do   
-       
-    wahl=$(kdialog --title "K+Radio" --radiolist "$station" 1 $station_status_1 2 $station_status_2 3 $station_status_3)
+ wahl=$(kdialog --title "K+Radio" --radiolist "$station" ${sender1[*]} ${sender2[*]} ${sender3[*]})
     
-    onoff=$?
-    
+ onoff=$?
+ #echo $onoff    
     if [ $onoff -eq 1 ]; then
     killall mpg123
     exit
     elif [ $onoff -eq 0 ]; then
     killall mpg123
+    sender1[2]=off
+    sender2[2]=off
+    sender3[2]=off
     fi
     
     
-    
-    
-    if [ "$wahl" = 3 ]; then
-    radiostation "http://swr-swr3-live.cast.addradio.de/swr/swr3/live/mp3/128/stream.mp3" 
-    station_status_3="SWR_3 on"
-    station_status_2="Dot_FM_8042 off"
-    station_status_1="Dot_FM_8072 off"
+   if [ "$wahl" = 3 ]; then
+    radiostation 3
+    sender3[2]=on
     elif [ "$wahl" = 2 ]; then
-    radiostation http://relay.181.fm:8042
-    station_status_3="SWR_3 off"
-    station_status_2="Dot_FM_8042 on"
-    station_status_1="Dot_FM_8072 off"
+    radiostation 2
+    sender2[2]=on
     elif [ "$wahl" = 1 ]; then
-    radiostation http://relay.181.fm:8072
-    station_status_3="SWR_3 off"
-    station_status_2="Dot_FM_8042 off"
-    station_status_1="Dot_FM_8072 on"
+    radiostation 1 
+    sender1[2]=on
     fi
-
-    done
+             
+done 
+   
+exit
